py_projects/prj_01_graphflow_py/graphflow/requirements.txt
fastapi==0.121.3
pandas==2.3.3
passlib==1.7.4
pydantic==2.12.4
pydantic_settings==2.12.0
PyJWT==2.10.1
python-dotenv==1.2.1
python_jose==3.5.0
SQLAlchemy==2.0.44
starlette==0.50.0